requirements.txt
aiofiles==24.1.0

aiohttp==3.12.12

python-dotenv==1.1.1
